Investor in das Programm eingebunden

See merge request inf_se2_sose25_startupx/inf_se2_sose25_startupx!28

+++ src/main/java/org/hbrs/se2/project/startupx/dtos/UnterstuetzungsangebotDTO.java
@@ -11,5 +11,5 @@ public class UnterstuetzungsangebotDTO {
     private Long id;
     private Long investor;
     private Long startup;
-    private Long betrag;
+    private Double betrag;
 }
